Money Cast
We 3 students live in one house, we buy things together,
such like foods, water and other things. So we need to
take notes of each people how much money we have used
and in the end, we have to count money for each person.
This help us to take notes and count easier.